Implemented FixtureGroup abstraction. FixtureGroup extends Fixtures to provide a publicly  accessible ArrayList of member Fixtures, as well as providing implementations of the interface methods of Fixture which transparently enable a FixtureGroup to be manipulated as if it were a single Fixture.
All Traits have been re-written to be FixtureGroup aware, allowing Traits to be applied to FixtureGroups of the appropriate type. Trait methods detect the type of their parent object and respond appropriately, iterating over the member Fixtures of the group if they detect a FixtureGroup.

ShowCOntrol now contains testing code for FixtureGroups in the form of a HueRotateBehavior applied to a group of two Fixtures, updating both simultaneously.


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@505 1c6c7da2-4a7c-0410-9345-fc53533f7dce