bug:64
afaict, the problem isn't that we're *not* getting the value on the midi
send/rcv path, it's that we're getting the wrong value.  in other words, we're
getting the old value, so we're rebroadcasting that.  rather than not
broadcasting the new value.

work around this in two ways:
- by hand rebroadcast the osc value immediately upon receipt
- then mark that as pending, and ignore the first received midi controller change that matches the same channel/number pair

limited testing, i'd like more complete testing in a real environment (with
multiple touchosc controllers)


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@525 1c6c7da2-4a7c-0410-9345-fc53533f7dce