i found yet another set of changes from the playa, this was from looking at the code deployed on controller C.
this concerns reading the panel index from a config file, where A, B,
and C are indexed by 0, 1, and 2 respectively.

the config file only contains a single value, no key.

i am not, however, checking in the exact changes i found on the playa
machine, for two reasons:

1) i can't actually get that code to work consistently.  loadStrings()
is relative to the data dir, which isn't easy to edit, since it ends
up in a jar in a compiled app.  the playa change tries to get around
this with "../panel_index.txt".  that works for me when running from
within processing, but not as a standalone app.  i can get it to work
for a standalone app by changing it to "panel_index.txt" and putting
it in the data dir, but then it's hard to change.  my solution foor
now is to move it to /opt/syzygryd/etc/panel_index.txt, which is
somewhat consistent with the config file (sequencer.properties) for
the sequencer.  [perhaps the filesystem setup ends up being different
on windows, and that's how it worked on the playa.]

2) the edits on controller C look wrong.  specifically, it looks like
once again changes were not applied correctly, and much (or all?) of
svn r715 was undone.  since the motivation for that rev was to work
around set switching issues, perhaps some continuing instability on
the playa was caused by those edits never having been properly
applied.

so *all* i'm doing here is the config file edit.

note that i really don't like this particular change.  it was perhaps
fine for the playa, so i do want to check this in, but it has some
issues:

- i'd rather a generic key/value pair config file, like we have for
  the sequencer.  this would make it easy to add other stuff, like not
  having to hardcode the ip addr.
- there are no concept of default values.
- a file that allowed comments would be nice.
- the code is quite fragile.  loadStrings() will return null if the
  file is found, which will cause a following line to throw a
  NullPointerException.  this totally freezes that app, and when
  running standalone, you can't even ESC out, you have to `pkill
  java`.


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@722 1c6c7da2-4a7c-0410-9345-fc53533f7dce