wow, hiding the cursor is a lot simpler than i had thought
i guess we don't have to worry about all of (from svn r727):
   bin/hide-cursor.sh
   bin/show-cursor.sh
   etc/rc.local.fragement
and maybe not even:
   etc/sudoers.fragment
if i really want to remove those abilities

for now keep them all in the repo, but i have for now commented out
the /etc/rc.local edit on controller3

so right now the code is hardcoded for the touchscreen.  next step is
to expose that as a boolean property.


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@728 1c6c7da2-4a7c-0410-9345-fc53533f7dce