mostly logging improvements.
also comment out some unused stuff, esp. the osc path from the
switcher to live (the return path is used).

and for now comment out ActionSetPlay.toString() b/c it's causing a
NPE, and there's now super.toString() that can be used instead.

and for now turn on debug logging, but we probably don't want to leave
it like that.  i should make that a property.

also some name changes, and comment additions, clarifiying what's
going on.  but some of it is just asking questions, and contemplating
about other possible cases of badness that maybe should be handled but
aren't.

basically i'm kind of still in the middle of stuff, but i want to
check it in now to try running on the syzyputer.


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@792 1c6c7da2-4a7c-0410-9345-fc53533f7dce