add another level of logging, verbose. for now, checked in is that debug is enabled, verbose is disabled. these really should be properties. or better yet, preprocessed away when not applicable.
add thread id to log header
log header ought to use StringBuffer (or StringBuilder) instead of String

add a quit live option to the web ui.  there really should be a
confirmation popup for this (as well as for quitting the switcher, if
that ought to even be exposed at all).

implement a more robust quitting of live, and invoke it in exceptional
conditions.  pressing enter to get the "This action will stop
audio. Proceed?" popup to go away doesn't work.  best to just hope
that avoiding it (via stopping with space if applicable) is
successful.

we need to be waiting, however, for the space to be registered as
stopped before proceeding.  in this checkin, that's an arbitrary wait
sharing code from ActionRunner.doStop() will come next.

the whole quit path is a bit harder than i had anticipated.  see
control flow of and comments within ProcessUtils.doLiveQuit() for more
details.


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@799 1c6c7da2-4a7c-0410-9345-fc53533f7dce