variable load timeout.  rather than 60 sec, 90 sec if we have to start live, 30 sec otherwise.
i hate the idea of an arbitrary sleep between sets, but i'm reluctant
to totally get rid of it without knowing why it was added initially.
for now decrease from 3 to 1 sec.

comment out sending enter to live.  this doesn't work when we maybe
need it, and likely causes problems in other situations.

what's the point of sending esc to live?  it's only called from the
web ui.  for now keep it.

get rid of some commented out stuff that was for initial testing
only.


git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@815 1c6c7da2-4a7c-0410-9345-fc53533f7dce